requirements.txt
chromadb==1.5.1
groq==1.0.0
huggingface_hub==1.4.1
pdfplumber==0.11.9
sentence-transformers==5.2.3
streamlit==1.54.0
torch==2.10.0
transformers==5.2.0
python-dotenv==1.2.1
